api/requirements.txt
Flask-RESTful~=0.3.8
Flask-SQLAlchemy~=2.5.1
PyJWT[crypto]~=2.0.1
gunicorn~=20.1.0
psycopg2~=2.8.6
pytest~=6.2.3
